+++ src/pathogena/models.py
@@ -43,7 +43,7 @@ class UploadBase(BaseModel):
         default=None, description="ISO 3166-2 principal subdivision"
     )
     district: str = Field(default="", description="Granular location")
-    specimen_organism: Literal["mycobacteria", "sars-cov-2", ""] = Field(
+    specimen_organism: Literal["mycobacteria", "sars-cov-2"] = Field(
         default="mycobacteria", description="Target specimen organism scientific name"
     )
     host_organism: str | None = Field(